third_party/blink/web_tests/external/wpt/css/css-cascade/scope-nesting-expected.txt

This is a testharness.js-based test.
[FAIL] Insert a nested style rule within @scope, &
  assert_equals: expected "rgb(255, 0, 0)" but got "rgb(0, 128, 0)"
Harness: the test ran to completion.
